utilised go routines and WaitGroup to asynchronously get prices

+++ api/responses.go
@@ -1,6 +1,5 @@
 package api
 
-
 // created struct to map CEX JSON response
 type CEXResponse struct {
 	Timestamp             string  `json:"timestamp"`
@@ -10,7 +9,7 @@ type CEXResponse struct {
 	Volume                string  `json:"volume"`
 	Volume30D             string  `json:"volume30d"`
 	Bid                   float64 `json:"bid"`
-	Ask                   int     `json:"ask"`
+	Ask                   float64 `json:"ask"`
 	PriceChange           string  `json:"priceChange"`
 	PriceChangePercentage string  `json:"priceChangePercentage"`
 	Pair                  string  `json:"pair"`
